requirements.txt
torch
torchvision
torchaudio
transformers
datasets<4.0.0
pytesseract
easyocr
opencv-python
Pillow
pandas
numpy
scikit-learn
matplotlib
tqdm
pyyaml
